FullTrottle/requirement.txt
asgiref==3.2.10
Django==3.1
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.11.1
mongoengine==0.20.0
pkg-resources==0.0.0
pymongo==3.11.0
pytz==2020.1
sqlparse==0.3.1
